chore: put all the activities, classes & interfaces in one separate subpackage except MainActivity
