+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:theme="@style/Theme.Planetze35"
         tools:targetApi="31">
         <activity
-            android:name=".EcoTrackerHomepageActivity"
+            android:name=".EcoTrackerAllParts.EcoTrackerHomepageActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -25,34 +25,34 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".EcoTrackerDailyActivityHub"
+            android:name=".EcoTrackerAllParts.EcoTrackerDailyActivityHub"
             android:exported="false" />
         <activity
-            android:name=".TransportationActivity"
+            android:name=".EcoTrackerAllParts.TransportationActivity"
             android:exported="false" />
         <activity
-            android:name=".FoodActivity"
+            android:name=".EcoTrackerAllParts.FoodActivity"
             android:exported="false" />
         <activity
-            android:name=".EnergyBillsActivity"
+            android:name=".EcoTrackerAllParts.EnergyBillsActivity"
             android:exported="false" />
         <activity
-            android:name=".ConsumptionEcoTrackerActivity"
+            android:name=".EcoTrackerAllParts.ConsumptionEcoTrackerActivity"
             android:exported="false" />
         <activity
-            android:name=".ViewListActivity"
+            android:name=".EcoTrackerAllParts.ViewListActivity"
             android:exported="false" />
         <activity
-            android:name=".SelectDateActionsActivity"
+            android:name=".EcoTrackerAllParts.SelectDateActionsActivity"
             android:exported="false" />
         <activity
-            android:name=".DatePickerActivity"
+            android:name=".EcoTrackerAllParts.DatePickerActivity"
             android:exported="false" />
         <activity
-            android:name=".RecommendationsActivity"
+            android:name=".EcoTrackerAllParts.RecommendationsActivity"
             android:exported="false" />
         <activity
-            android:name=".BaseRecommendationsActivity"
+            android:name=".EcoTrackerAllParts.BaseRecommendationsActivity"
             android:exported="false" />
         <activity
             android:name = ".MainActivity"
