+++ app/src/main/java/com/example/planetze35/EcoTrackerAllParts/ActivityItem.java
@@ -1,4 +1,4 @@
-package com.example.planetze35;
+package com.example.planetze35.EcoTrackerAllParts;
 
 public class ActivityItem {
     private String id;
