+++ app/src/main/java/com/example/planetze35/EcoTrackerAllParts/BaseRecommendationsActivity.java
@@ -1,4 +1,4 @@
-package com.example.planetze35;
+package com.example.planetze35.EcoTrackerAllParts;
 
 import android.Manifest;
 import android.app.NotificationChannel;
@@ -15,6 +15,8 @@ import androidx.core.app.ActivityCompat;
 import androidx.core.content.ContextCompat;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.example.planetze35.R;
 
 import java.util.ArrayList;
