+++ app/src/main/java/com/example/planetze35/EcoTrackerAllParts/EcoTrackerDailyActivityHub.java
@@ -1,4 +1,4 @@
-package com.example.planetze35;
+package com.example.planetze35.EcoTrackerAllParts;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -9,6 +9,8 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.graphics.Insets;
 import androidx.core.view.ViewCompat;
 import androidx.core.view.WindowInsetsCompat;
+
+import com.example.planetze35.R;
 
 public class EcoTrackerDailyActivityHub extends AppCompatActivity {
