+++ app/src/main/java/com/example/planetze35/EcoTrackerAllParts/EnergyBills.java
@@ -1,4 +1,4 @@
-package com.example.planetze35;
+package com.example.planetze35.EcoTrackerAllParts;
 
 import java.util.HashMap;
 import java.util.Map;
