+++ app/src/main/java/com/example/planetze35/EcoTrackerAllParts/FetchUserDailyActivities.java
@@ -1,4 +1,4 @@
-package com.example.planetze35;
+package com.example.planetze35.EcoTrackerAllParts;
 
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
