+++ app/src/main/java/com/example/planetze35/EcoTrackerAllParts/FilterHelper.java
@@ -1,4 +1,4 @@
-package com.example.planetze35;
+package com.example.planetze35.EcoTrackerAllParts;
 
 import android.content.Context;
