+++ app/src/main/java/com/example/planetze35/EcoTrackerAllParts/Habit.java
@@ -1,4 +1,4 @@
-package com.example.planetze35;
+package com.example.planetze35.EcoTrackerAllParts;
 
 public class Habit {
     private final String name;
