+++ app/src/main/res/layout/activity_date_picker.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/bg_1"
-    tools:context=".DatePickerActivity">
+    tools:context=".EcoTrackerAllParts.DatePickerActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
