+++ app/src/main/res/layout/activity_eco_tracker_daily_hub.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/eco_tracker_activity_bg"
-    tools:context=".EcoTrackerDailyActivityHub">
+    tools:context=".EcoTrackerAllParts.EcoTrackerDailyActivityHub">
 
     <GridLayout
         android:id="@+id/grid_buttons"
