+++ app/src/main/res/layout/activity_eco_tracker_homepage.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/bg_1"
-    tools:context=".EcoTrackerHomepageActivity">
+    tools:context=".EcoTrackerAllParts.EcoTrackerHomepageActivity">
 
     <LinearLayout
         android:orientation="vertical"
