+++ app/src/main/res/layout/activity_select_date_actions.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/bg_1"
-    tools:context=".SelectDateActionsActivity">
+    tools:context=".EcoTrackerAllParts.SelectDateActionsActivity">
 
     <!-- GridLayout for buttons -->
     <GridLayout
