+++ app/src/main/res/layout/activity_view_list.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/bg_1"
-    tools:context=".ViewListActivity">
+    tools:context=".EcoTrackerAllParts.ViewListActivity">
 
     <!-- Header TextView -->
     <TextView
